+++ src/entities/Ground.h
@@ -3,13 +3,16 @@
 
 #include "../util/Utils.h"
 
+#include <SDL2/SDL.h>
+#include <SDL2/SDL_image.h>
+
 class Ground
 {
 private:
     SDL_Renderer* renderer;
     SDL_Texture* texture;
     SDL_Rect frame;
-    Uint32 last_update_time;
+    float last_update_time;
     SDL_FPoint* position;
     float velocity;
     SDL_FRect* bounding_box;
@@ -18,7 +21,7 @@ public:
     Ground(SDL_Renderer* renderer, float velocity);
     virtual ~Ground();
 
-    void update(Uint32 current_time);
+    void update(float current_time);
     void render();
     inline SDL_FPoint* getPosition() const { return position; }
     inline SDL_FRect* getBoundingBox() const { return bounding_box; }
